fix: KV cache expansion for duplicated attention layers, RYS startup log, server custom_logs
- Add add_kv_layer() to llama_kv_cache for extra KV cache slots (RYS dup layers)
- Add get_k_remapped()/get_v_remapped() to separate cache lookup from hparams queries
- Thread cache_il parameter through build_attn(inp_kv) for remapped KV cache access
- Expand KV cache in create_memory for QWEN35 when RYS is enabled
- Route duplicated attention layers to separate cache slots in qwen35.cpp
- Add RYS startup log in llama_init_from_model for all tools
- Wire llama_set_custom_logs in server.cpp
- Fix format warnings in qwen35/qwen3next RYS logs
- Update architecture doc: correct arch from qwen3next to qwen35
- Update instructions doc: revised test commands, server/MTP notes

+++ src/models/qwen3next.cpp
@@ -148,7 +148,7 @@ llama_model_qwen3next::graph::graph(const llama_model & model, const llm_graph_p
         }
         LLAMA_LOG_INFO("%s: RYS layer looping enabled: layers %d-%d repeated %d times, effective layers: %zu (original: %d)\n",
                        __func__, cparams.loop_layer_start, cparams.loop_layer_stop,
-                       cparams.loop_count, layer_schedule.size(), n_layer);
+                       cparams.loop_count, layer_schedule.size(), (int) n_layer);
     } else if (cparams.custom_logs) {
         LLAMA_LOG_INFO("%s: RYS layer looping disabled (loop_count=%d, loop_layer_start=%d, loop_layer_stop=%d)\n",
                        __func__, cparams.loop_count, cparams.loop_layer_start, cparams.loop_layer_stop);
